+++ src/neuroconv/tools/roiextractors/__init__.py
@@ -2,9 +2,12 @@ from .roiextractors import (
     get_nwb_imaging_metadata,
     add_devices,
     add_imaging_plane,
+    add_image_segmentation,
+    add_plane_segmentation,
     add_two_photon_series,
     add_epochs,
     write_imaging,
     get_nwb_segmentation_metadata,
+    add_summary_images,
     write_segmentation,
 )